apps/backend/processing-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
celery==5.4.0
redis==5.0.7
aio-pika==9.4.1
h5py==3.11.0
numpy==1.26.4
onnxruntime==1.18.0
Pillow==10.3.0
scikit-learn==1.5.0
boto3==1.34.131
httpx==0.27.0
python-multipart==0.0.9
pyjwt>=2.8.0
python-dotenv>=1.0.0
# MCAP / rosbag2 support
mcap>=1.1.0
mcap-ros2-support>=0.5.0
